Remove SDL Schema generation test and move SDL generator into commons
