+++ commons/src/main/java/io/ballerina/stdlib/graphql/commons/schema/types/DefaultDirective.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.commons.types;
+package io.ballerina.stdlib.graphql.commons.schema.types;
 
 import java.util.Arrays;
 import java.util.List;
