+++ commons/src/main/java/io/ballerina/stdlib/graphql/commons/schema/types/Type.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.commons.types;
+package io.ballerina.stdlib.graphql.commons.schema.types;
 
 import java.io.Serializable;
 import java.util.ArrayList;
