--- compiler-plugin-tests/src/test/resources/ballerina_sources/generator_tests/24_sdl_file_generation_with_multiple_services1/schema3.graphql
@@ -1,4 +1,0 @@
-type Query {
-  name: String!
-  age: Int!
-}
