--- compiler-plugin-tests/src/test/resources/ballerina_sources/generator_tests/25_sdl_file_generation_with_multiple_services2/Ballerina.toml
@@ -1,4 +1,0 @@
-[package]
-org = "graphql_test"
-name = "test_package"
-version = "0.1.0"
