--- compiler-plugin-tests/src/test/resources/ballerina_sources/generator_tests/25_sdl_file_generation_with_multiple_services2/schema1.graphql
@@ -1,8 +1,0 @@
-type Query {
-  profile: Person!
-}
-
-type Person {
-  name: String!
-  age: Int!
-}
