--- compiler-plugin-tests/src/test/resources/ballerina_sources/generator_tests/25_sdl_file_generation_with_multiple_services2/schema2.graphql
@@ -1,7 +1,0 @@
-type Query {
-  school: School!
-}
-
-type School {
-  name(name: String!): String!
-}
