--- compiler-plugin-tests/src/test/resources/ballerina_sources/generator_tests/26_sdl_file_generation_with_custom_scalar_types/schema.graphql
@@ -1,14 +1,0 @@
-type Query {
-  height(height: Decimal!): Decimal!
-}
-
-# The `Decimal` type corresponds to a subset of IEEE 754-2008 128-bit decimal (radix 10) floating point numbers
-scalar Decimal
-
-type Mutation {
-  singleFileUpload(file: Upload!): String!
-  multipleFileUpload(files: [Upload!]!): String!
-}
-
-# The `Upload` type represents file upload in a GraphQL schema
-scalar Upload
