+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/diagnostics/CompilationDiagnostic.java
@@ -64,8 +64,7 @@ public enum CompilationDiagnostic {
                                              DiagnosticSeverity.ERROR),
     INVALID_ANONYMOUS_FIELD_TYPE(DiagnosticMessage.ERROR_130, DiagnosticCode.GRAPHQL_130, DiagnosticSeverity.ERROR),
     INVALID_ANONYMOUS_INPUT_TYPE(DiagnosticMessage.ERROR_131, DiagnosticCode.GRAPHQL_131, DiagnosticSeverity.ERROR),
-    INVALID_RETURN_TYPE_CLASS(DiagnosticMessage.ERROR_132, DiagnosticCode.GRAPHQL_132, DiagnosticSeverity.ERROR),
-    SDL_FILE_GENERATION_FAILED(DiagnosticMessage.WARNING_133, DiagnosticCode.GRAPHQL_133, DiagnosticSeverity.WARNING);
+    INVALID_RETURN_TYPE_CLASS(DiagnosticMessage.ERROR_132, DiagnosticCode.GRAPHQL_132, DiagnosticSeverity.ERROR);
 
     private final String diagnostic;
     private final String diagnosticCode;
