+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/diagnostics/DiagnosticCode.java
@@ -53,6 +53,5 @@ public enum DiagnosticCode {
     GRAPHQL_129,
     GRAPHQL_130,
     GRAPHQL_131,
-    GRAPHQL_132,
-    GRAPHQL_133
+    GRAPHQL_132
 }
