+++ docs/spec/spec.md
@@ -464,8 +464,6 @@ The GraphQL schema is generated by analyzing the Ballerina service attached to t
 
 When an incompatible type is used inside a GraphQL service, a compilation error will be thrown.
 
-The generated GraphQL schema is exported as a Schema Definition Language(SDL) file by the GraphQL package. For the Ballerina projects, the schema file is written into `target` directory and for the single bal files, the schema is written into the same directory where the command is executed.
-
 ### 3.1 Root Types
 Root types are a special set of types in a GraphQL schema. These types are associated with an operation, which can be done on the GraphQL scheme. There are three root types.
